11.2.Fixing N+1 With a Join?{A Smarter Join?}
Let's remove the join and rely on the EXTRA_LAZY solution.

A Smarter Join?
Yes, this will mean that we will once again have 27 queries. If you don't like that, there is another solution: you could make the JOIN query smarter - it would look like this:

// src/Repository/BigFootSightingRepository.php
public function findLatestQueryBuilder(int $maxResults): QueryBuilder
{
    return $this->createQueryBuilder('big_foot_sighting')
        ->leftJoin('big_foot_sighting.comments', 'comments')
        ->groupBy('big_foot_sighting.id')
        ->addSelect('COUNT(comments.id) as comment_count')
        ->setMaxResults($maxResults)
        ->orderBy('big_foot_sighting.createdAt', 'DESC');
}
The key is that instead of selecting all the comment data... which we don't need... this selects only the count. It gets the exact data we need, in one query. From a performance standpoint, it's probably the perfect solution.
But... it has a downside: complexity. Instead of returning an array of BigFootSighting objects, this will return an array of... arrays... where each has a 0 key that is the BigFootSighting object and a comment_count key with the count. It's just... a bit weird to deal with. For example, the template would need to be updated to take this into account:

{% for sightingData in sightings %}
    {% set sighting = sightingData.0 %}
    {% set commentCount = sightingData.comment_count %}

    {# ... #}
        {{ sighting.title }}

        {{ commentCount }}
    {# ... #}
{% endfor %}
And... because of the pagination that this app is using... the new query would actually produce an error. So let's keep things how they are now. If the extra queries ever become a real problem on production, then we can think about spending time improving this. Sometimes profiling is about knowing what not to fix... because it may not be worth it.
Next, if you were surprised that we didn't see any evidence of the network request that the homepage is making to render the SymfonyCasts repository info from GitHub, that's because the homepage is more complex than it might seem. Let's use a cool "Profile all" feature to see all requests that the homepage makes.